fixed a bug in the camera zoom
it turns out I need to account for matt's starting position, so I did

theres another bug with dialogue that is definitely on matt's side, but I have no intention of fixing that today

the issue is when an NPC has dialogue that has no lines for Matt, the dialogue works fine once and doesnt reset. This is definitely due to one of the variables in Areamatt not resetting correctly, which can be fixed by sending a signal from the NPC Collision code. However, I am not sure which variable is not being reset, so I need to do a lot of debugging to figure it out.